+++ frontend/src/App.tsx
@@ -1,10 +1,11 @@
 import './App.css'
-import NavBar from "./Components/NavBar/NavBar.tsx";
+import NavBar from "./Components/NavBar.tsx";
 import {Outlet} from "react-router";
 import "react-toastify/dist/ReactToastify.css"  // styles for the toast
 import {ToastContainer} from "react-toastify";
 import {UserProvider} from "../Context/useAuth.tsx";
-// import Hero from "./Components/Hero/Hero.tsx";
+import store from './Redux/store.ts';
+import {Provider} from "react-redux";
 
 function App() {
     
@@ -12,9 +13,11 @@ function App() {
   return (
     <>
         <UserProvider>
-            <NavBar/>
-            <Outlet/>
-            <ToastContainer/>
+            <Provider store={store}>
+                <NavBar/>
+                <Outlet/>
+                <ToastContainer/>
+            </Provider>
         </UserProvider>
     </>
   );
