+++ frontend/src/Components/CardPortfolio.tsx
@@ -1,0 +1,20 @@
+import {SyntheticEvent} from "react";
+import DeletePortfolio from "./DeletePortfolio.tsx";
+import {Link} from "react-router-dom";
+import {PortfolioGet} from "../Types/PortfolioGet";
+
+interface Props {
+    portfolioValue: PortfolioGet;
+    onPortfolioDelete: (e: SyntheticEvent) => void;
+}
+
+const CardPortfolio = ({portfolioValue, onPortfolioDelete}: Props) => {
+    return (
+        <div className="flex flex-col w-full p-8 space-y-4 text-center rounded-lg shadow-lg md:w-1/3">
+            <Link to={`/company/${portfolioValue.symbol}/company-profile`} className="pt-6 text-xl font-bold">{portfolioValue.symbol}</Link>
+            <DeletePortfolio onPortfolioDelete={onPortfolioDelete} portfolioValue={portfolioValue.symbol}/>
+        </div>
+    );
+};
+
+export default CardPortfolio;
