--- frontend/src/Components/CompanyDashboard/CompanyDashboard.tsx
@@ -1,29 +1,0 @@
-import {Outlet} from "react-router-dom";
-import {ReactNode} from "react";
-
-interface Props {
-    children: ReactNode;
-    ticker: string;
-};
-
-const CompanyDashboard = ({children, ticker}: Props) => {
-    // we use React.FC here to include default children props
-    return (
-        <div className="relative md:ml-64 bg-blueGray-100 w-full">
-            <div className="relative pt-20 pb-32 bg-lightBlue-500">
-                <div className="px-4 md:px-6 mx-auto w-full">
-                    <div>
-                        <div className="flex flex-wrap">
-                            {children}
-                        </div>
-                        <div className="flex flex-wrap">
-                            {<Outlet context={ticker}/>}
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    );
-};
-
-export default CompanyDashboard;
