+++ frontend/src/Components/Search.tsx
@@ -1,0 +1,30 @@
+import {ChangeEvent, SyntheticEvent} from "react";
+
+interface Props {
+    search: string | undefined;
+    onSearchSubmit: (e: SyntheticEvent) => void;
+    handleSearchChange: (e: ChangeEvent<HTMLInputElement>) => void;
+}
+
+const Search = ({search, onSearchSubmit, handleSearchChange}: Props) => {
+    return (
+        <section className="relative bg-gray-100">
+            <div className="max-w-4xl mx-auto p-6 space-y-6">
+                <form
+                    className="form relative flex flex-col w-full p-10 space-y-4 bg-darkBlue rounded-lg md:flex-row md:space-y-0 md:space-x-3"
+                    onSubmit={onSearchSubmit}
+                >
+                    <input
+                        className="flex-1 p-3 rounded-lg placeholder-black focus:outline-none bg-white"
+                        id="search-input"
+                        placeholder="Search companies"
+                        value={search}
+                        onChange={handleSearchChange}
+                    ></input>
+                </form>
+            </div>
+        </section>
+    );
+}
+
+export default Search;
