+++ frontend/src/Components/StockComment/StockCommentForm.tsx
