+++ frontend/src/Pages/HomePage.tsx
@@ -1,0 +1,11 @@
+import Hero from "../Components/Hero.tsx";
+
+export const HomePage = () => {
+    return (
+        <div>
+            <Hero/>
+        </div>
+    );
+};
+
+export default HomePage;
