--- frontend/src/Pages/HomePage/HomePage.tsx
@@ -1,11 +1,0 @@
-import Hero from "../../Components/Hero/Hero.tsx";
-
-export const HomePage = () => {
-    return (
-        <div>
-            <Hero/>
-        </div>
-    );
-};
-
-export default HomePage;
